Refactor: video player 컴포넌트 분리
