+++ src/app/[roomId]/page.tsx
@@ -37,6 +37,7 @@ import useAddPlaylist from '@/hooks/use-add-playlist';
 import useDeletePlaylist from '@/hooks/use-delete-playlist';
 import Chat from '@/components/chat';
 import ParticipantsList from '@/components/participants-list';
+import VideoPlayer from '@/components/video-player';
 import { hasVideoEditPermission } from '@/service/user';
 
 const RoomPage = ({ params: { roomId } }: { params: { roomId: string } }) => {
@@ -75,20 +76,12 @@ const RoomPage = ({ params: { roomId } }: { params: { roomId: string } }) => {
   const { data: userInfo } = useGetUserInfo({ roomCode });
   const { data: playlist = [] } = useGetPlaylist({ roomCode });
   const { data: roomDetailInfo } = useGetRoomDetailInfo({ roomCode });
-  const { data: videoTitleInfo } = useGetVideoTitleInfo({ roomCode });
-  const { data: videoSyncInfo } = useGetVideoSyncInfo({ roomCode });
   const { mutate: changeUserRole } = useChangeRole();
   const { mutate: addPlaylist } = useAddPlaylist();
   const { mutate: deletePlaylist } = useDeletePlaylist();
-  const { mutate: playNextVideo } = usePlayNextVideo();
   const [chatValue, setChatValue] = useState('');
   const participantsList = participants?.[0]?.participants;
   const playlistInfo = playlist?.[0]?.playlist;
-
-  const playerRef = useRef<YouTubePlayer | null>(null);
-  const isFirstVideoPlayRef = useRef<boolean>(false);
-  const [isPlayerReady, setIsPlayerReady] = useState(false);
-  const [isMuted, setIsMuted] = useState<boolean>(false);
 
   const queryClient = useQueryClient();
   const { register, handleSubmit, watch, reset } = useForm<TYoutubeUrlPayload>({
@@ -134,157 +127,6 @@ const RoomPage = ({ params: { roomId } }: { params: { roomId: string } }) => {
       );
     }
   }, [roomCode, isSuccessOfGetVideoInfo, videoInfo, addPlaylist, queryClient]);
-
-  useEffect(() => {
-    const curPlayerState = playerRef.current?.getPlayerState();
-    const playerState: { [key: number]: string } = {
-      1: 'PLAY',
-      2: 'PAUSE',
-      0: 'END',
-    };
-
-    if (
-      isPlayerReady &&
-      playerRef.current &&
-      videoSyncInfo?.playerState !== playerState[curPlayerState]
-    ) {
-      if (
-        videoSyncInfo?.playerState === 'PLAY' &&
-        !isFirstVideoPlayRef.current
-      ) {
-        playerRef.current.mute();
-        playerRef.current.playVideo();
-        isFirstVideoPlayRef.current = true;
-      } else if (
-        videoSyncInfo?.playerState === 'PLAY' &&
-        isFirstVideoPlayRef.current
-      ) {
-        if (isMuted) playerRef.current.mute();
-        playerRef.current.playVideo();
-      } else if (videoSyncInfo?.playerState === 'PAUSE') {
-        playerRef.current.pauseVideo();
-      } else if (videoSyncInfo?.playerState === 'END') {
-        playerRef.current.stopVideo();
-        setIsPlayerReady(false);
-        setIsMuted(playerRef.current.isMuted());
-        playerRef.current = null;
-
-        queryClient.setQueryData<TVideoSyncInfo>(['videoSyncInfo', roomCode], {
-          videoId: null,
-          playerState: 'END',
-          playerCurrentTime: 0,
-          playerRate: 1,
-          videoNumber: null,
-        });
-
-        queryClient.setQueryData<TVideoTitleInfo>(
-          ['videoTitleInfo', roomCode],
-          {
-            videoTitle: null,
-            channelTitle: null,
-          }
-        );
-      }
-    }
-
-    if (isPlayerReady && playerRef.current) {
-      const playerCurrentTime = playerRef.current?.getCurrentTime();
-
-      if (
-        videoSyncInfo?.playerCurrentTime &&
-        Math.abs(playerCurrentTime - videoSyncInfo?.playerCurrentTime) > 0.6
-      ) {
-        playerRef.current?.seekTo(videoSyncInfo?.playerCurrentTime);
-        if (videoSyncInfo.playerState === 'PAUSE') {
-          playerRef.current?.pauseVideo();
-        }
-      }
-
-      const playerCurrentRate = playerRef.current?.getPlaybackRate();
-
-      if (
-        videoSyncInfo?.playerRate &&
-        videoSyncInfo?.playerRate !== playerCurrentRate
-      ) {
-        playerRef.current?.setPlaybackRate(videoSyncInfo?.playerRate);
-      }
-    }
-  }, [videoSyncInfo, isPlayerReady, isMuted, roomCode, queryClient]);
-
-  const handleReadyState = (event: YouTubeEvent) => {
-    playerRef.current = event.target;
-    setIsPlayerReady(true);
-  };
-
-  const handleNextVideoButton = () => {
-    playNextVideo({ videoNumber: playlistInfo[0].videoNumber });
-  };
-  // 음소거 설정 상태 저장하기
-  const handlePlayerStateChange = (event: YouTubeEvent) => {
-    const newPlayerState = event.data;
-    const playerState: { [key: number]: string } = {
-      1: 'PLAY',
-      2: 'PAUSE',
-      0: 'END',
-    };
-    // video 상태를 변경시킨 사용자가 아닌, 상태 변화를 전달받은 사용자는
-    // 다시 상태를 변경하지 않도록 하기 위한 조건문
-    if (playerState[newPlayerState] === videoSyncInfo?.playerState) return;
-
-    // 권한 없는 사용자인 경우 alert 띄우기
-    if (
-      !userHasVideoEditPermission &&
-      playerState[newPlayerState] === 'PLAY' &&
-      playerState[newPlayerState] !== videoSyncInfo?.playerState
-    ) {
-      playerRef.current?.seekTo(videoSyncInfo?.playerCurrentTime);
-      playerRef.current?.pauseVideo();
-      return;
-    } else if (
-      !userHasVideoEditPermission &&
-      playerState[newPlayerState] === 'PAUSE' &&
-      playerState[newPlayerState] !== videoSyncInfo?.playerState
-    ) {
-      playerRef.current?.seekTo(videoSyncInfo?.playerCurrentTime);
-      playerRef.current?.playVideo();
-      return;
-    }
-
-    if (playerState[newPlayerState] === 'PLAY') {
-      sendVideoPlayerState({
-        roomCode: roomCode,
-        playerState: 'PLAY',
-        playerCurrentTime: event.target.getCurrentTime(),
-        playerRate: event.target.getPlaybackRate(),
-      });
-    } else if (playerState[newPlayerState] === 'PAUSE') {
-      sendVideoPlayerState({
-        roomCode: roomCode,
-        playerState: 'PAUSE',
-        playerCurrentTime: event.target.getCurrentTime(),
-        playerRate: event.target.getPlaybackRate(),
-      });
-    }
-  };
-
-  const handlePlayerRateChange = (event: YouTubeEvent) => {
-    const newPlayerRate = event.target.getPlaybackRate();
-
-    if (newPlayerRate === videoSyncInfo?.playerRate) return;
-
-    // 권한 없는 사용자인 경우 alert 띄우기
-    if (!userHasVideoEditPermission) {
-      playerRef.current?.setPlaybackRate(videoSyncInfo?.playerRate);
-      return;
-    }
-
-    sendVideoPlayerState({
-      roomCode: roomCode,
-      playerState: 'RATE',
-      playerCurrentTime: event.target.getCurrentTime(),
-      playerRate: newPlayerRate,
-    });
-  };
 
   if (isLoading)
     return (
@@ -341,69 +183,12 @@ const RoomPage = ({ params: { roomId } }: { params: { roomId: string } }) => {
       <NavBar isHomePage={false} />
 
       <div className="flex w-full h-auto justify-center items-start px-40 gap-4">
-        <div className="flex flex-col gap-2">
-          {!videoSyncInfo?.videoId && (
-            <div className="flex items-center justify-center w-videoWidth h-videoHeight bg-emptyPlaylist">
-              <div className="flex flex-col items-center">
-                <p className="text-default-400">재생목록이 비었습니다</p>
-                <p className="text-default-300">{`There are no videos in the room's playlist`}</p>
-              </div>
-            </div>
-          )}
-          {videoSyncInfo?.videoId && (
-            <YouTube
-              key={videoSyncInfo?.videoNumber}
-              videoId={videoSyncInfo?.videoId}
-              opts={{
-                width: 680,
-                height: 480,
-                playerVars: {
-                  autoplay: 0,
-                  disablekb: 1,
-                  rel: 0,
-                },
-              }}
-              onReady={handleReadyState}
-              onStateChange={handlePlayerStateChange}
-              onPlaybackRateChange={handlePlayerRateChange}
-            />
-          )}
-          <div className="flex gap-2 items-center justify-between">
-            {videoTitleInfo?.videoTitle && (
-              <div className="flex flex-col gap-1">
-                <p className="text-sm text-emerald-500">현재 재생중인 영상</p>
-                <p className="text-lg">{videoTitleInfo?.videoTitle}</p>
-                <p className="text-xs text-neutral-400">
-                  {videoTitleInfo?.channelTitle}
-                </p>
-              </div>
-            )}
-            {!videoTitleInfo?.videoTitle && (
-              <div className="flex flex-col gap-1">
-                <p className="text-sm text-emerald-500">
-                  현재 재생중인 영상이 없습니다
-                </p>
-              </div>
-            )}
-            <Button
-              size="sm"
-              variant="light"
-              disabled={
-                !userHasVideoEditPermission || playlistInfo?.length === 0
-              }
-              onPress={() => handleNextVideoButton()}
-            >
-              <Icon
-                name="playNextVideo"
-                className={`size-5 ${
-                  !userHasVideoEditPermission || playlistInfo?.length === 0
-                    ? 'text-neutral-700'
-                    : 'text-emerald-500'
-                }`}
-              />
-            </Button>
-          </div>
-        </div>
+        <VideoPlayer
+          roomCode={roomCode}
+          userInfo={userInfo}
+          sendVideoPlayerState={sendVideoPlayerState}
+          playlistInfo={playlistInfo}
+        />
         <div className="flex flex-col w-80 gap-2">
           <div className="w-full min-h-10 p-3 border-small rounded-small border-default-200 dark:border-default-100">
             <div className="flex justify-between items-center">
